Move from JSON -> .ts
