+++ tsconfig.json
@@ -3,7 +3,6 @@
     "target": "es2019",
     "module": "commonjs",
     "moduleResolution": "node",
-    "resolveJsonModule": true,
     "esModuleInterop": true,
     "strict": true,
     "rootDir": ".",
